暂时写成左键下棋，右键把棋子移除(后面根据需求慢慢改ba)。左键点击已经下好的棋子可以显示棋子信息

+++ src/ck/ckrc/erie/warchess/Director.java
@@ -17,7 +17,7 @@ import java.net.URL;
 import java.util.Optional;
 
 public class Director {
-    public static final double width=900,height=700;
+    public static final int width=900,height=700;
     private Stage stage;
     private GameScene gameScene=new GameScene();
     private GameOver overScene=new GameOver();
